Add(III): Fix
