+++ src/Main.java
@@ -1,5 +1,4 @@
 import figures.Line;
-import figures.Square;
 
 import java.util.Scanner;
 
@@ -7,14 +6,11 @@ public class Main {
     public static void main(String[] args) {
         int option = 0;
         Scanner scanner = new Scanner(System.in);
-        Line line = new Line();
-        Square square = new Square();
         do {
+            Line line = new Line();
             System.out.println("------ Figures System ------");
             System.out.println("1. Create Line");
             System.out.println("2. Print Line");
-            System.out.println("3. Create Square");
-            System.out.println("4. Print Square");
             System.out.println("3. Exit");
             option = scanner.nextInt();
             switch (option) {
@@ -27,20 +23,12 @@ public class Main {
                     line.drawLine();
                     break;
                 case 3:
-                    System.out.println("Please enter Square length: ");
-                    int squareLength = scanner.nextInt();
-                    square.setLength(squareLength);
-                    break;
-                case 4:
-                    square.drawSquare();
-                    break;
-                case 5:
                     System.out.println("------ Exiting System -------");
                     break;
                 default:
                     System.out.println("Please enter a valid option");
                     break;
             }
-        }while (option != 5);
+        }while (option != 3);
     }
 }
